Move parse_time_string to squid.utils
